Refactored and improved MotorControl
Used to be ControlMotor.
